requirements.txt
PIL==1.1.7
Django==1.2
django-auth-ldap==1.0.12
django-autoslug==1.4.1
django-extensions==0.7.1
django-treemenus==0.8.7
eav-django==1.3.4
html5lib==0.90
pisa==3.0.33
python-ldap==2.4.4
reportlab==2.5
wsgiref==0.1.2
